// src/app/interfaces/dailyReadings.interface.ts
import { Reading } from 'src/app/interfaces/reading.interface';

export interface DailyReadings {
    Feast: string;
    First_reading: string;
    First_readingsrc: string;
    Psalm: string;
    Psalmsrc: string;
    Second_reading: string;
    Second_readingsrc: string;
    Gospel: string;
    Gospelsrc: string;
    allelluia: string;
  };